Fix crash in StandardTokenizer if wchar_t is 32bit
StandardTokenizerImpl::getNextToken() uses a map to classify characters;
the map is 16bit, because this code was ported from Java and Java's
strings use UTF-16. But Lucene++ uses wchar_t, which may be 32bit on
some platforms (Unix, OS X), resulting in crashes when encountering
characters above U+FFFF.

Fix the crash by classifying anything outside of BMP as a character.
This is hardly ideal, but it prevents the crash and gets the job done
reasonably well (as StandardTokenizer is intended for European languages
in the first place). A proper fix would rework the map to correspond to
UTF-32 rather than UTF-16, or would port Unicode tokenizer from Lucene
3.1+.

Fixes #57.